requirements.txt
annotated-types==0.7.0
anyio==4.12.1
audioop-lts==0.2.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
colorama==0.4.6
comtypes==1.4.16
distro==1.9.0
flatbuffers==25.12.19
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
mpmath==1.3.0
numpy==2.4.2
onnxruntime==1.24.3
openai==2.26.0
packaging==26.0
piper-tts==1.4.1
playwright==1.58.0
protobuf==7.34.0
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==311
requests==2.32.5
sniffio==1.3.1
sounddevice==0.5.5
SpeechRecognition==3.14.5
srt==3.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
sympy==1.14.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
vosk==0.3.45
websockets==16.0
